Shrunk player down shrunk grid size
Shrunk the grid size so that the tiles fit the whole grid box. I think this isn't the best way to fix the problem I had but Its the only solution i was able to acheive. I editted the movement to adjust to the smaller map size and editted the feel.